+++ packages/video_player_videohole/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## NEXT
 
 * Adds compatibility with `http` 1.0 in example.
+* Removed unused proxy APIs.
 
 ## 0.5.8